requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.1
PyMuPDF==1.23.21
langchain-openai==0.0.5